Update README with new code file details
# HeadDirectionEncoding 
The new code is in ring-attractor-revision one.rar. For the project, the code ring_attractor.py is the main function and the implementation for Vergani and Huyck’s  RAN
Another new code is in ring-attractor-master-Revision one.rar. For the project, the file ring_attractorAdaptiveOld.py is the AdaptiveRAN and the file ring_attractorAssociationOld.py is the Pantelis et. al’s  RAN.